requirements.txt
anthropic>=0.39.0
mcp>=1.2.0
python-dotenv>=1.0.1
rich>=13.7.1
